refactor file name/head/doc
