+++ lmcache/v1/multiprocess/server.py
@@ -210,7 +210,7 @@ def _build_modules(
     lmcache_driven_cls: type[LMCacheDrivenTransferModule] = LMCacheDrivenTransferModule
     if mp_config.layerwise_batch > 0:
         # First Party
-        from lmcache.v1.multiprocess.modules.lmcache_layerwise_transfer import (
+        from lmcache.v1.multiprocess.modules.lmcache_driven_transfer_layerwise import (
             LMCacheLayerwiseTransferModule,
         )
